Refactor BERTrend (moved functions inside the class)
