+++ bertrend/BERTrend.py
@@ -2,6 +2,7 @@
 #  See AUTHORS.txt
 #  SPDX-License-Identifier: MPL-2.0
 #  This file is part of BERTrend.
+import os
 import pickle
 import shutil
 from collections import defaultdict
@@ -15,12 +16,14 @@ from loguru import logger
 from pandas import Timestamp
 from sentence_transformers import SentenceTransformer
 from sklearn.metrics.pairwise import cosine_similarity
+from tqdm import tqdm
 
 from bertrend import (
     MODELS_DIR,
     CACHE_PATH,
     BERTREND_DEFAULT_CONFIG_PATH,
     load_toml_config,
+    SIGNAL_EVOLUTION_DATA_DIR,
 )
 
 from bertrend.BERTopicModel import BERTopicModel
@@ -39,6 +42,9 @@ from bertrend.trend_analysis.weak_signals import (
     _initialize_new_topic,
     update_existing_topic,
     _apply_decay_to_inactive_topics,
+    _filter_data,
+    _is_rising_popularity,
+    _create_dataframes,
 )
 from bertrend.utils.data_loading import TEXT_COLUMN
 
@@ -450,7 +456,7 @@ class BERTrend:
         self.topic_last_popularity = topic_last_popularity
         self.topic_last_update = topic_last_update
 
-    def compute_popularity_values_and_thresholds(
+    def _compute_popularity_values_and_thresholds(
         self, window_size: int, current_date: Timestamp
     ) -> tuple[Timestamp, Timestamp, list, float, float]:
         """
@@ -491,6 +497,124 @@ class BERTrend:
             q1, q3 = 0, 0
 
         return window_start, window_end, all_popularity_values, q1, q3
+
+    def _classify_signals(
+        self,
+        window_start: pd.Timestamp,
+        window_end: pd.Timestamp,
+        q1: float,
+        q3: float,
+        rising_popularity_only: bool = True,
+        keep_documents: bool = True,
+    ) -> tuple[pd.DataFrame, pd.DataFrame, pd.DataFrame]:
+        """
+        Classify signals into weak signal and strong signal dataframes.
+
+        Args:
+            window_start (pd.Timestamp): The start timestamp of the window.
+            window_end (pd.Timestamp): The end timestamp of the window.
+            q1 (float): The 10th percentile of popularity values.
+            q3 (float): The 50th percentile of popularity values.
+            rising_popularity_only (bool): Whether to consider only rising popularity topics as weak signals.
+            keep_documents (bool): Whether to keep track of the documents or not.
+
+        Returns:
+            Tuple[pd.DataFrame, pd.DataFrame, pd.DataFrame]:
+                - noise_topics_df: DataFrame containing noise topics.
+                - weak_signal_topics_df: DataFrame containing weak signal topics.
+                - strong_signal_topics_df: DataFrame containing strong signal topics.
+        """
+        noise_topics = []
+        weak_signal_topics = []
+        strong_signal_topics = []
+
+        sorted_topics = sorted(self.topic_sizes.items(), key=lambda x: x[0])
+
+        for topic, data in sorted_topics:
+            filtered_data = _filter_data(data, window_end, keep_documents)
+            if not filtered_data["Timestamps"]:
+                continue
+
+            window_popularities = [
+                (timestamp, popularity)
+                for timestamp, popularity in zip(
+                    filtered_data["Timestamps"], filtered_data["Popularity"]
+                )
+                if window_start <= timestamp <= window_end
+            ]
+
+            if window_popularities:
+                latest_timestamp, latest_popularity = window_popularities[-1]
+                docs_count = (
+                    filtered_data["Docs_Count"][-1]
+                    if filtered_data["Docs_Count"]
+                    else 0
+                )
+                paragraphs_count = (
+                    filtered_data["Paragraphs_Count"][-1]
+                    if filtered_data["Paragraphs_Count"]
+                    else 0
+                )
+                source_diversity = (
+                    filtered_data["Source_Diversity"][-1]
+                    if filtered_data["Source_Diversity"]
+                    else 0
+                )
+
+                topic_data = (
+                    topic,
+                    latest_popularity,
+                    latest_timestamp,
+                    docs_count,
+                    paragraphs_count,
+                    source_diversity,
+                    filtered_data,
+                )
+
+                if latest_popularity < q1:
+                    noise_topics.append(topic_data)
+                elif q1 <= latest_popularity <= q3:
+                    if rising_popularity_only:
+                        if _is_rising_popularity(filtered_data, latest_timestamp):
+                            weak_signal_topics.append(topic_data)
+                        else:
+                            noise_topics.append(topic_data)
+                    else:
+                        weak_signal_topics.append(topic_data)
+                else:
+                    strong_signal_topics.append(topic_data)
+
+        return _create_dataframes(
+            noise_topics, weak_signal_topics, strong_signal_topics, keep_documents
+        )
+
+    def classify_signals(
+        self, window_size: int, current_date: Timestamp
+    ) -> tuple[pd.DataFrame, pd.DataFrame, pd.DataFrame]:
+        """
+        Classify signals into weak signal and strong signal dataframes for the considered time window.
+
+        Args:
+            window_size (int): The retrospective window size in days.
+            current_date (datetime): The current date selected by the user.
+
+
+        Returns:
+            Tuple[pd.DataFrame, pd.DataFrame, pd.DataFrame]:
+                - noise_topics_df: DataFrame containing noise topics.
+                - weak_signal_topics_df: DataFrame containing weak signal topics.
+                - strong_signal_topics_df: DataFrame containing strong signal topics.
+        """
+        # Compute threshold values
+        window_start, window_end, all_popularity_values, q1, q3 = (
+            self._compute_popularity_values_and_thresholds(window_size, current_date)
+        )
+
+        # Classify signals
+        noise_topics_df, weak_signal_topics_df, strong_signal_topics_df = (
+            self._classify_signals(window_start, window_end, q1, q3)
+        )
+        return noise_topics_df, weak_signal_topics_df, strong_signal_topics_df
 
     def save_models(self, models_path: Path = MODELS_DIR):
         if models_path.exists():
@@ -571,6 +695,63 @@ class BERTrend:
         bertrend.topic_models = topic_models
 
         return bertrend
+
+    def save_signal_evolution_data(
+        self,
+        window_size: int,
+        start_timestamp: pd.Timestamp,
+        end_timestamp: pd.Timestamp,
+    ) -> Path:
+        save_path = SIGNAL_EVOLUTION_DATA_DIR / f"retrospective_{window_size}_days"
+        os.makedirs(save_path, exist_ok=True)
+
+        q1_values, q3_values, timestamps_over_time = [], [], []
+        noise_dfs, weak_signal_dfs, strong_signal_dfs = [], [], []
+
+        for current_timestamp in tqdm(
+            pd.date_range(
+                start=start_timestamp,
+                end=end_timestamp,
+                freq=pd.Timedelta(days=self.config["granularity"]),
+            ),
+            desc="Processing timestamps",
+        ):
+            window_start, window_end, all_popularity_values, q1, q3 = (
+                self._compute_popularity_values_and_thresholds(
+                    window_size, current_timestamp
+                )
+            )
+
+            noise_df, weak_signal_df, strong_signal_df = self._classify_signals(
+                window_start, window_end, q1, q3, keep_documents=False
+            )
+
+            noise_dfs.append(noise_df)
+            weak_signal_dfs.append(weak_signal_df)
+            strong_signal_dfs.append(strong_signal_df)
+
+            timestamps_over_time.append(current_timestamp)
+
+        # Save the grouped dataframes
+        with open(save_path / "noise_dfs_over_time.pkl", "wb") as f:
+            pickle.dump(noise_dfs, f)
+        with open(save_path / "weak_signal_dfs_over_time.pkl", "wb") as f:
+            pickle.dump(weak_signal_dfs, f)
+        with open(save_path / "strong_signal_dfs_over_time.pkl", "wb") as f:
+            pickle.dump(strong_signal_dfs, f)
+
+        # Save the metadata
+        with open(save_path / "metadata.pkl", "wb") as f:
+            metadata = {
+                "window_size": window_size,
+                "granularity": self.config["granularity"],
+                "timestamps": timestamps_over_time,
+                "q1_values": q1_values,
+                "q3_values": q3_values,
+            }
+            pickle.dump(metadata, f)
+
+        return save_path
 
 
 def _preprocess_model(
